% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fake_seqs.R
\docType{data}
\name{fake_seqs}
\alias{fake_seqs}
\title{Example FASTA dataset}
\format{
A character vector of length 6 (FASTA text lines).
}
\source{
Simulated data created by Marcel Ferreira.
}
\usage{
fake_seqs
}
\description{
A small example FASTA file with 30 sequences, included for demonstration
purposes in the package.
}
\keyword{datasets}
