% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/length_dna.R
\name{length_dna}
\alias{length_dna}
\title{Calcule o tamanho da sequencia de DNA}
\usage{
length_dna(tbl, keepSequence = TRUE)
}
\arguments{
\item{tbl}{Uma tabela (data.frame ou tibble) que contenha a coluna 'sequence'}

\item{keepSequence}{Um valor logico para manter (default) a coluna que contem as sequencias}
}
\value{
Um `tbl` contendo a coluna BPs que informa a quantidade de
}
\description{
Esta função recebe uma tabela contendo a coluna 'sequence' e calcula o total de bases de DNA.
}
